update to version 3.8.2

+++ tests/key-material-set-dtls.c
@@ -15,12 +15,11 @@
  * General Public License for more details.
  *
  * You should have received a copy of the GNU General Public License
- * along with GnuTLS; if not, write to the Free Software Foundation,
- * Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA
+ * along with GnuTLS.  If not, see <https://www.gnu.org/licenses/>.
  */
 
 #ifdef HAVE_CONFIG_H
-#include <config.h>
+#include "config.h"
 #endif
 
 #include <stdio.h>
@@ -71,8 +70,10 @@ static void terminate(void)
 {
 	int status = 0;
 
-	kill(child, SIGTERM);
-	wait(&status);
+	if (child) {
+		kill(child, SIGTERM);
+		wait(&status);
+	}
 	exit(1);
 }
 
@@ -99,13 +100,14 @@ static void client(int fd)
 
 	/* Initialize TLS session
 	 */
-	gnutls_init(&session, GNUTLS_CLIENT|GNUTLS_DATAGRAM);
+	gnutls_init(&session, GNUTLS_CLIENT | GNUTLS_DATAGRAM);
 	gnutls_record_set_timeout(session, 10000);
 
 	/* Use default priorities */
-	ret = gnutls_priority_set_direct(session,
-				   "NONE:+VERS-DTLS1.0:+AES-128-CBC:+SHA1:+SIGN-ALL:+COMP-NULL:+ANON-DH:+ANON-ECDH:+CURVE-ALL",
-				   &err);
+	ret = gnutls_priority_set_direct(
+		session,
+		"NONE:+VERS-DTLS1.0:+AES-128-CBC:+SHA1:+SIGN-ALL:+COMP-NULL:+ANON-DH:+ANON-ECDH:+CURVE-ALL",
+		&err);
 	if (ret < 0) {
 		fail("client: priority set failed (%s): %s\n",
 		     gnutls_strerror(ret), err);
@@ -122,12 +124,11 @@ static void client(int fd)
 	 */
 	do {
 		ret = gnutls_handshake(session);
-	}
-	while (ret < 0 && gnutls_error_is_fatal(ret) == 0);
+	} while (ret < 0 && gnutls_error_is_fatal(ret) == 0);
 
 	if (ret < 0) {
 		fail("client: Handshake failed: %s\n", strerror(ret));
-		terminate();
+		exit(1);
 	} else {
 		if (debug)
 			success("client: Handshake was completed\n");
@@ -135,38 +136,42 @@ static void client(int fd)
 
 	if (debug)
 		success("client: TLS version is: %s\n",
-			gnutls_protocol_get_name
-			(gnutls_protocol_get_version(session)));
+			gnutls_protocol_get_name(
+				gnutls_protocol_get_version(session)));
 
 	ret = gnutls_cipher_get(session);
 	if (ret != GNUTLS_CIPHER_AES_128_CBC) {
-		fprintf(stderr, "negotiated unexpected cipher: %s\n", gnutls_cipher_get_name(ret));
-		terminate();
+		fprintf(stderr, "negotiated unexpected cipher: %s\n",
+			gnutls_cipher_get_name(ret));
+		exit(1);
 	}
 
 	ret = gnutls_mac_get(session);
 	if (ret != GNUTLS_MAC_SHA1) {
-		fprintf(stderr, "negotiated unexpected mac: %s\n", gnutls_mac_get_name(ret));
-		terminate();
+		fprintf(stderr, "negotiated unexpected mac: %s\n",
+			gnutls_mac_get_name(ret));
+		exit(1);
 	}
 
 	/* save state */
-	ret = gnutls_record_get_state(session, 0, NULL, NULL, NULL, wseq_number);
+	ret = gnutls_record_get_state(session, 0, NULL, NULL, NULL,
+				      wseq_number);
 	if (ret < 0) {
 		fprintf(stderr, "error in %d\n", __LINE__);
 		gnutls_perror(ret);
-		terminate();
+		exit(1);
 	}
 
-	ret = gnutls_record_get_state(session, 1, NULL, NULL, NULL, rseq_number);
+	ret = gnutls_record_get_state(session, 1, NULL, NULL, NULL,
+				      rseq_number);
 	if (ret < 0) {
 		fprintf(stderr, "error in %d\n", __LINE__);
 		gnutls_perror(ret);
-		terminate();
+		exit(1);
 	}
 
 	/* skip past the sliding window */
-	for (i=0;i<96;i++) {
+	for (i = 0; i < 96; i++) {
 		ret = gnutls_record_send(session, "hello", 5);
 		if (ret < 0) {
 			fail("gnutls_record_send: %s\n", gnutls_strerror(ret));
@@ -187,14 +192,14 @@ static void client(int fd)
 	if (ret < 0) {
 		fprintf(stderr, "error in %d\n", __LINE__);
 		gnutls_perror(ret);
-		terminate();
+		exit(1);
 	}
 
 	ret = gnutls_record_set_state(session, 1, rseq_number);
 	if (ret < 0) {
 		fprintf(stderr, "error in %d\n", __LINE__);
 		gnutls_perror(ret);
-		terminate();
+		exit(1);
 	}
 
 	ret = gnutls_record_send(session, "ping", 4);
@@ -231,8 +236,7 @@ static void server(int fd)
 	unsigned char rseq_number[8];
 	unsigned char wseq_number[8];
 	char buf[128];
-	const gnutls_datum_t p3 =
-	    { (unsigned char *) pkcs3, strlen(pkcs3) };
+	const gnutls_datum_t p3 = { (unsigned char *)pkcs3, strlen(pkcs3) };
 
 	/* this must be called once in the program
 	 */
@@ -248,14 +252,14 @@ static void server(int fd)
 	gnutls_dh_params_import_pkcs3(dh_params, &p3, GNUTLS_X509_FMT_PEM);
 	gnutls_anon_set_server_dh_params(anoncred, dh_params);
 
-	gnutls_init(&session, GNUTLS_SERVER|GNUTLS_DATAGRAM);
+	gnutls_init(&session, GNUTLS_SERVER | GNUTLS_DATAGRAM);
 	gnutls_record_set_timeout(session, 10000);
 
 	/* avoid calling all the priority functions, since the defaults
 	 * are adequate.
 	 */
-	ret = gnutls_priority_set_direct(session,
-				   "NORMAL:+VERS-DTLS1.0:+ANON-DH:+ANON-ECDH", NULL);
+	ret = gnutls_priority_set_direct(
+		session, "NORMAL:+VERS-DTLS1.0:+ANON-DH:+ANON-ECDH", NULL);
 	if (ret < 0) {
 		fail("server: priority set failed (%s)\n\n",
 		     gnutls_strerror(ret));
@@ -268,8 +272,7 @@ static void server(int fd)
 
 	do {
 		ret = gnutls_handshake(session);
-	}
-	while (ret < 0 && gnutls_error_is_fatal(ret) == 0);
+	} while (ret < 0 && gnutls_error_is_fatal(ret) == 0);
 	if (ret < 0) {
 		close(fd);
 		gnutls_deinit(session);
@@ -282,16 +285,18 @@ static void server(int fd)
 
 	if (debug)
 		success("server: TLS version is: %s\n",
-			gnutls_protocol_get_name
-			(gnutls_protocol_get_version(session)));
+			gnutls_protocol_get_name(
+				gnutls_protocol_get_version(session)));
 
 	/* save state */
-	ret = gnutls_record_get_state(session, 0, NULL, NULL, NULL, wseq_number);
+	ret = gnutls_record_get_state(session, 0, NULL, NULL, NULL,
+				      wseq_number);
 	if (ret < 0) {
 		fail("error in %d\n", __LINE__);
 	}
 
-	ret = gnutls_record_get_state(session, 1, NULL, NULL, NULL, rseq_number);
+	ret = gnutls_record_get_state(session, 1, NULL, NULL, NULL,
+				      rseq_number);
 	if (ret < 0) {
 		fail("error in %d\n", __LINE__);
 	}
@@ -308,7 +313,7 @@ static void server(int fd)
 
 			ret = gnutls_record_send(session, buf, ret);
 		}
-	} while(ret > 0);
+	} while (ret > 0);
 
 	if (ret < 0) {
 		fail("error: %s\n", gnutls_strerror(ret));
@@ -389,4 +394,4 @@ void doit(void)
 	start();
 }
 
-#endif				/* _WIN32 */
+#endif /* _WIN32 */
